requirements.txt
alembic==1.4.3
aniso8601==9.0.1
anyio==3.6.2
application==1.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
ecdsa==0.18.0
fastapi==0.92.0
Flask==2.2.3
flask-marshmallow==0.9.0
Flask-Migrate==2.3.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.1
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
jwt==1.3.1
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==2.13.6
marshmallow-sqlalchemy==0.15.0
mysqlclient==1.3.13
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-editor==1.0.4
python-jose==3.3.0
pytz==2022.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.2.13
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
Werkzeug==2.2.3
zipp==3.14.0
